+++ CHANGES.md
@@ -73,6 +73,7 @@ now for questions and discussions about development. [GH-820]
   * Properly handle Go build failing in a directory with multiple packages
     [GH-676] [GH-904]
   * Make cppcheck recognise C++ header files [GH-909]
+  * Don’t run phpcs on empty buffers [GH-907]
 
 [buttercup]: https://github.com/jorgenschaefer/emacs-buttercup